Develop

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -1,14 +1,24 @@
 <?php
 
 namespace App\Http\Controllers\Admin;
+
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
+use App\Models\Product;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
 class CategoryController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('permission:categories_list')->only(['index']);
+        $this->middleware('permission:categories_create')->only(['create', 'store']);
+        $this->middleware('permission:categories_detail')->only(['show']);
+        $this->middleware('permission:categories_update')->only(['edit', 'update']);
+        $this->middleware('permission:categories_delete')->only(['destroy']);
+    }
     /**
      * Display a listing of the resource.
      */
@@ -98,7 +108,8 @@ class CategoryController extends Controller
         if ($request->hasFile('image')) {
             $data['image'] = $request->file('image')->store('categories', 'public');
             if ($category->image && file_exists(storage_path('app/public/' . $category->image))) {
-                Storage::delete('public/' . $category->image);            }
+                Storage::delete('public/' . $category->image);
+            }
         }
         $category->update($data);
 
@@ -111,20 +122,18 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::findOrFail($id);
-        // $hash = Product::where('category_id', $id)->exists();
-        // if (!$hash) {
-        $category->delete();
-        $catefories = Category::all();
-        foreach ($catefories as $x) {
-            if ($x->parent_id == $id) {
-                $x->delete();
+        $hash = Product::where('category_id', $id)->exists();
+        if (!$hash) {
+            $category->delete();
+            $catefories = Category::all();
+            foreach ($catefories as $x) {
+                if ($x->parent_id == $id) {
+                    $x->delete();
+                }
             }
+            return redirect()->route('categories.index')->with('success', 'Xoá danh mục thành công !');
+        } else {
+            return redirect()->route('admin.categories.index')->with('error', 'Vui lòng chuyển các sản phẩm sang danh mục khác để tiền hành xoá danh mục này.');
         }
-        return redirect()->route('categories.index')->with('success', 'Xoá danh mục thành công !');
-
-        // }else{
-        //     return redirect()->route('admin.categories.index')->with('error', 'Vui lòng chuyển các sản phẩm sang danh mục khác để tiền hành xoá danh mục này.');
-
-        // }
     }
 }
